[UserList] Refactor: split dialog code to separate file
